requirements.txt
# ORAM-Integrated PyTorch Training Baseline
# CSE239A Thesis Project - Caden Roberts

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# ORAM Implementation
pyoram>=0.3.0

# Cryptography (PyORAM backend)
cryptography>=41.0.0
pycryptodome>=3.19.0

# Profiling and Analysis
memory-profiler>=0.61.0
psutil>=5.9.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
